comment skeleton and post/homescreen header formatting
I went ahead and put the basic code for displaying and writing new comments on the post screen, Its mostly just a skeleton as it has no interaction with the database. I also fixed an issue with the headers for the home and post screens from my last commit and added a back button on the post screen header.